FastAPI/requirements.txt
fastapi
uvicorn
tensorflow-cpu
pillow
scikit-learn
pydantic
protobuf
typing-extensions
python-multipart
pydicom
keras